Splitting source code into multiple files
