+++ apps/parser/CMakeLists.txt
@@ -1,5 +1,5 @@
 set(CMAKE_CXX_STANDARD 20)
 
 find_package(Boost REQUIRED)
-add_executable(parser parser.cpp)
+add_executable(parser parser.cpp domain.cpp)
 target_include_directories(parser PRIVATE ${Boost_INCLUDE_DIR})
