+++ apps/parser/config.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <boost/spirit/home/x3.hpp>
+
+namespace client { namespace parser
+{
+    namespace x3 = boost::spirit::x3;
+
+    using iterator_type = std::string::const_iterator;
+}}
